refactor(organisms): merge SfCarousel
