--- packages/vue/src/components/organisms/SfCarousel/SfCarousel.html
@@ -1,30 +1,0 @@
-<div class="sf-carousel">
-  <div class="sf-carousel__controls">
-    <!--@slot slot for icon moving to the previous item -->
-    <slot name="prev" v-bind="{ go: () => go('prev') }">
-        <SfArrow
-          class="sf-arrow--long"
-          aria-label="previous"
-          @click="go('prev')"
-        />
-    </slot>
-    <!--@slot slot for icon moving to the next item -->
-    <slot name="next" v-bind="{ go: () => go('next') }">
-        <SfArrow
-          class="sf-arrow--long sf-arrow--right"
-          aria-label="next"
-          @click="go('next')"
-        />
-    </slot>
-  </div>
-  <div class="sf-carousel__wrapper">
-    <div class="glide" ref="glide">
-      <div class="glide__track" data-glide-el="track">
-        <ul class="glide__slides sf-carousel__slides">
-          <!--@slot default slot for SfCarouselItem tags -->
-          <slot />
-        </ul>
-      </div>
-    </div>
-  </div>
-</div>
